updated filestructure. fixed location bug. added basic funcionality including not-found page. basic style updates.
