+++ legislative-awareness/package.json
@@ -21,6 +21,7 @@
     "react-dom": "^17.0.2",
     "react-dropdown": "^1.9.2",
     "react-helmet": "^6.1.0",
+    "react-resizable": "^3.0.4",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "react-social-icons": "^5.6.1",
