+++ legislative-awareness/public/index.html
@@ -2,12 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/logoGov-removebg.png" />
+    <link rel="icon" href="%PUBLIC_URL%/icon-removebg.png"/>
+    <!-- <link rel="stylesheet" href="%SRC_URL%/App.css"> -->
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="Web site containing legislative information in New York State"
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logoGov-removebg.png" />
     <!--
