+++ legislative-awareness/src/App.css
@@ -1,3 +1,374 @@
+/* GENERAL STYLING */
+.footer {
+  margin-top: 3em;
+}
+
+.navigation {
+  border-bottom: solid black;
+  border-radius: 20px;
+  position: fixed sticky;
+}
+
+/* HEADER */
+
+.underline {
+  border-bottom: solid black 3px;
+  border-radius: 10%;
+  padding-left: 1em;
+  padding-bottom: 0.3em;
+}
+
+header {
+  color: black;
+  text-transform: capitalize;
+  text-align: center;
+  padding: 3em;
+  border-bottom: solid black 5px;
+  border-radius: 3em;
+}
+
+.textBlock {
+  color: black;
+  text-align: left;
+  padding-top: 2em;
+  padding-left: 3em;
+  padding-right: 3em;
+  padding-bottom: 2em;
+  max-width: 50%;
+}
+
+.textBlock.medium {
+  padding-top: 6em !important;
+}
+
+/* Confused yet? Good */
+.textBlock.medium.small {
+  padding-top: 4.7em !important;
+}
+
+.smallBlock {
+  text-align: left;
+  padding-top: 2em;
+  padding-left: 3em;
+  padding-right: 3em;
+  padding-bottom: 2em;
+  max-width: 33%;
+}
+
+.split {
+  display: flex;
+  flex-direction: row;
+}
+
+/* HOME STYLING */
+
+.grid {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  /* flex-wrap: wrap; */
+}
+
+.cell {
+  /* AESTHETICS:  */
+  border: 1px solid black;
+  border-radius: 15px;
+  padding: 7em;
+  text-align: center;
+  margin: 0.6em;
+  background-color: gray;
+  cursor: pointer;
+  flex-shrink: 1;
+  opacity: 1;
+
+  /* GRID CELL:   */
+
+  display: flex;
+  transform: scale(1);
+  transition-duration: 0.6s;
+}
+
+.cell:hover {
+  opacity: 20;
+  transform: scale(1.02);
+  /* transform: rotate(2deg); */
+  transform-origin: center;
+  transition-duration: 0.8s;
+}
+
+.cell.normal {
+  flex-grow: 1;
+  color: black;
+}
+
+.cell.beegboi {
+  flex-grow: 2;
+  color: black;
+}
+
+.cell.data {
+  background-image: url("./components/assets/dataImg2.png");
+  backdrop-filter: blur(4px);
+  background-position: -20em;
+}
+
+.cell.local {
+  background-image: url("./components/assets/localImg.jpg");
+}
+
+.cell.information {
+  background-image: url("./components/assets/informationImg.jpg");
+  backdrop-filter: blur(4px);
+}
+
+.cell.about {
+  background-image: url("./components/assets/aboutImg.jpg");
+  background-position: -14em -25.5em;
+
+  backdrop-filter: blur(4px);
+}
+
+.item {
+  border: black 1px;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  width: 150%;
+}
+.item.text {
+  border: black 1px;
+  padding: 2em;
+}
+
+.item.img {
+  border: black 1px;
+  /* padding: 2em; */
+}
+
+.white {
+  color: white;
+  transition-duration: 0.8s;
+}
+
+.white:hover {
+  color: dodgerblue;
+  transition-duration: 0.6s;
+}
+
+.black {
+  color: black;
+  transition-duration: 0.8s;
+}
+
+.black:hover {
+  color: dodgerblue;
+  transition-duration: 0.6s;
+}
+
+.link {
+  color: black;
+  text-decoration: none;
+}
+
+/* ADDRESS SEARCH STYLING */
+
+.search-icon {
+  position: absolute;
+  top: 6px;
+  left: 8px;
+  width: 14px;
+}
+
+.clear-icon {
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  width: 10px;
+  cursor: pointer;
+  /* <!-- visibility: hidden; --> */
+}
+
+.toggle-buttons {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.searchbar {
+  width: 16.2rem;
+  margin-top: 0.5rem;
+}
+
+
+.searchbar.input {
+  transform: scale(1);
+  transition: 0.2s;
+}
+
+.searchbar.input:hover {
+  transform: scale(1.2);
+  transition: 0.2s;
+}
+
+.social-media > li {
+  display: inline-block;
+  padding-left: 0.5rem;
+}
+
+.social-icon {
+  width: 25px !important;
+  height: 25px !important;
+}
+
+/* For the reps: */
+
+/* .box-grid {
+
+} */
+
+.box {
+  padding: 1em;
+  background-image: linear-gradient(
+    rgba(138, 182, 169, 0.5),
+    rgba(255, 255, 255, 0)
+  );
+  /* border:2px solid rgb(170, 241, 232); */
+  border-radius: 20px;
+  margin: 0.6em;
+
+  transform: scale(1);
+  transition-duration: 0.6s;
+}
+
+.box:hover {
+  transform: scale(1.02);
+  transition-duration: 0.6s;
+}
+
+/* For the layout: */
+
+.splitItem {
+  padding: 1em;
+  color: black;
+}
+
+.splitItem.Beeg {
+  /* float: left; */
+  max-height: 45em;
+  overflow-y: scroll;
+  max-width: 96%;
+  padding-left: 35em;
+}
+
+.splitItem.Beeg::-webkit-scrollbar {
+  display: none;
+}
+
+.splitItem.Smol {
+  /* float: left; */
+  position: fixed;
+  max-width: 33%;
+}
+
+.headerText {
+  padding-top: 2em;
+  padding-right: 2em;
+  padding-left: 2em;
+  padding-bottom: 0.3em;
+}
+
+.descriptionText {
+  padding-left: 5em;
+}
+
+/* .box {
+  color: #fff;
+  align-items: center;
+  text-align: center;
+  max-width: flex;
+  margin: 30px auto;
+  overflow: auto;
+  min-height: flex;
+  border: 1px solid white;
+  padding: 1em;
+  border-radius: 5px;
+} */
+
+.list {
+  align-content: left;
+}
+
+.container {
+  padding: 1em;
+  margin: 0 auto;
+  text-align: center;
+}
+
+.rep-image {
+  height: 100px;
+  width: 100px;
+}
+
+ul {
+  text-align: left;
+}
+
+ul > ul {
+  background-color: #f4f4f4;
+  padding: 2em;
+  border: 1px solid #fdfdfd;
+  margin-bottom: 1rem !important;
+}
+
+
+/* INFORMATION PAGE: */
+
+.center-header {
+  text-align: center;
+  align-content: center;
+  padding: 1em;
+  border-bottom:ridge black ;
+  /* border-radius: 10%; */
+}
+
+.thin {
+  margin: auto;
+  /* border: solid black; */
+  max-width: 70%;
+}
+
+.middle {
+  justify-content: center;
+}
+
+.centeredText {
+  text-align: center;
+  max-width:50%;
+  padding: 2em;
+  border-bottom: ridge black;
+}
+
+
+/* FOOTER: */
+
+/* needs work */
+.foot {
+  background: linear-gradient(
+    0deg,
+    rgba(67, 67, 67, 1) 0%,
+    rgba(255, 255, 255, 1) 55%
+  );
+}
+
+.proximity-dropdown-div {
+  width: 50%;
+  float: left;
+  padding: 20px;
+}
+
+
+/* Old Styling: */
 .page {
   display: flex;
   flex-direction: column;
@@ -116,4 +487,44 @@ td {
   border: 1px solid black;
   border-radius: 25px;
   margin: 10px;
+}
+
+
+
+#center {
+  /* background-image: linear-gradient(rgb(167, 255, 140), rgba(255,255,255, 0)); */
+  background-image: url("components/assets/404NotFound.jpg");
+ 
+  background-size: cover;
+  
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: auto;
+  padding-top:5em;
+  width: 100%;
+  padding-bottom: 40em;
+  /* padding-bottom: 30em; */
+}
+
+.smolButton {
+  height:3em;
+  width: 9em;
+  border-radius: 5px;
+
+  padding:8px;
+  text-align: center;
+
+  background:linear-gradient(rgba(255, 255, 255, 0),rgba(138, 182, 169, 0.5));
+  transform: scale(1);
+  transition: 0.5s;
+}
+
+.smolButton:hover {
+  transform: scale(1.2);
+  background:rgba(138, 182, 169, 0.5);
+  transition: 0.3s;
+  cursor:pointer;
 }
