+++ legislative-awareness/src/components/Footer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./style.css"
+// import "./style.css"
 
 function Footer() {
   return (
