+++ legislative-awareness/src/components/Header.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom"
-import "./style.css"
+// import "./style.css"
 
 export default function Header() {
