+++ legislative-awareness/src/components/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./style.css";
+// import "./style.css";
 import { Link } from 'react-router-dom';
 
 import Header from './Header'
