+++ legislative-awareness/src/components/index.js
@@ -5,7 +5,7 @@ export { default as Home } from "./Home";
 export { default as Local } from "./Local";
 export { default as Data } from "./Data";
 export { default as Information } from "./Information";
-
+export { NotFound } from "./NotFound";
 
 // export { default as About } from "./About";
 // export { default as Contact } from "./Contact";
